requirements.txt
diff-match-patch==20121119
Django==2.0.5
django-filter==1.1.0
django-import-export==1.0.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.3
psycopg2==2.7.4
pytz==2018.4
PyYAML==3.12
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0